web: refactor: decompose forge_hooks/trigger.rs into submodules
Split the 2226-line trigger.rs into payloads/installation/fanout/approval/commands with a parse_comment_event normalizer collapsing the per-forge comment branches; behavior preserved. Part of #479.
